refactor outputs manager
