+++ crates/pera-core/src/authority/authority_per_epoch_store.rs
@@ -69,10 +69,11 @@ use crate::consensus_handler::{
 use crate::consensus_manager::ConsensusManager;
 use crate::dwallet_mpc;
 use crate::dwallet_mpc::authority_name_to_party_id;
+use crate::dwallet_mpc::batches_manager::DWalletMPCBatchesManager;
 use crate::dwallet_mpc::mpc_manager::{
     DWalletMPCChannelMessage, DWalletMPCManager, DWalletMPCSender,
 };
-use crate::dwallet_mpc::mpc_outputs_manager::DWalletMPCOutputsManager;
+use crate::dwallet_mpc::mpc_outputs_verifier::DWalletMPCOutputsVerifier;
 use crate::epoch::epoch_metrics::EpochMetrics;
 use crate::epoch::randomness::{
     DkgStatus, RandomnessManager, RandomnessReporter, VersionedProcessedMessage,
@@ -346,7 +347,8 @@ pub struct AuthorityPerEpochStore {
     pub dwallet_mpc_sender: OnceCell<DWalletMPCSender>,
     /// State machine managing DWallet MPC outputs.
     /// This state machine only being used to store outputs and declare ones with quorum of votes as valid.
-    pub dwallet_mpc_outputs_manager: OnceCell<tokio::sync::Mutex<DWalletMPCOutputsManager>>,
+    pub dwallet_mpc_outputs_verifier: OnceCell<tokio::sync::Mutex<DWalletMPCOutputsVerifier>>,
+    pub dwallet_mpc_batches_manager: OnceCell<tokio::sync::Mutex<DWalletMPCBatchesManager>>,
 }
 
 /// AuthorityEpochTables contains tables that contain data that is only valid within an epoch.
@@ -863,8 +865,9 @@ impl AuthorityPerEpochStore {
             jwk_aggregator,
             randomness_manager: OnceCell::new(),
             randomness_reporter: OnceCell::new(),
-            dwallet_mpc_outputs_manager: OnceCell::new(),
+            dwallet_mpc_outputs_verifier: OnceCell::new(),
             dwallet_mpc_sender: OnceCell::new(),
+            dwallet_mpc_batches_manager: OnceCell::new(),
         });
         s.update_buffer_stake_metric();
         s
@@ -934,20 +937,35 @@ impl AuthorityPerEpochStore {
     }
 
     /// A function to initiate the [`DWalletMPCSender`] when a new epoch starts.
-    pub async fn set_dwallet_mpc_sender(&self, sender: DWalletMPCSender) -> PeraResult<()> {
+    pub fn set_dwallet_mpc_sender(&self, sender: DWalletMPCSender) -> PeraResult<()> {
         if self.dwallet_mpc_sender.set(sender).is_err() {
             error!("BUG: `set_dwallet_mpc_sender` called more than once; this should never happen");
         }
         Ok(())
     }
 
-    /// A function to initiate the Dwallet MPC outputs manager when a new epoch starts.
-    pub async fn set_dwallet_mpc_outputs_manager(
+    /// A function to initiate the [`DWalletMPCBatchesManager`] when a new epoch starts.
+    pub fn set_dwallet_mpc_batches_manager(
         &self,
-        manager: DWalletMPCOutputsManager,
+        batches_manager: DWalletMPCBatchesManager,
     ) -> PeraResult<()> {
         if self
-            .dwallet_mpc_outputs_manager
+            .dwallet_mpc_batches_manager
+            .set(tokio::sync::Mutex::new(batches_manager))
+            .is_err()
+        {
+            error!("BUG: `set_dwallet_mpc_batches_manager` called more than once; this should never happen");
+        }
+        Ok(())
+    }
+
+    /// A function to initiate the Dwallet MPC outputs verifier when a new epoch starts.
+    pub fn set_dwallet_mpc_outputs_manager(
+        &self,
+        manager: DWalletMPCOutputsVerifier,
+    ) -> PeraResult<()> {
+        if self
+            .dwallet_mpc_outputs_verifier
             .set(tokio::sync::Mutex::new(manager))
             .is_err()
         {
@@ -2636,10 +2654,24 @@ impl AuthorityPerEpochStore {
         Some(VerifiedSequencedConsensusTransaction(transaction))
     }
 
-    pub async fn get_dwallet_mpc_outputs_manager(
+    /// Return the [`DWalletMPCOutputsVerifier`].
+    /// Need to use a Mutex because the instance is initialized from a different thread.
+    pub async fn get_dwallet_mpc_outputs_verifier(
         &self,
-    ) -> PeraResult<tokio::sync::MutexGuard<DWalletMPCOutputsManager>> {
-        let dwallet_mpc_outputs_manager = self.dwallet_mpc_outputs_manager.get();
+    ) -> PeraResult<tokio::sync::MutexGuard<DWalletMPCOutputsVerifier>> {
+        let dwallet_mpc_outputs_manager = self.dwallet_mpc_outputs_verifier.get();
+        match dwallet_mpc_outputs_manager {
+            Some(dwallet_mpc_outputs_manager) => Ok(dwallet_mpc_outputs_manager.lock().await),
+            None => Err(PeraError::from(
+                "DWalletMPCOutputsManager is not initialized",
+            )),
+        }
+    }
+
+    pub async fn get_dwallet_mpc_batches_manager(
+        &self,
+    ) -> PeraResult<tokio::sync::MutexGuard<DWalletMPCBatchesManager>> {
+        let dwallet_mpc_outputs_manager = self.dwallet_mpc_batches_manager.get();
         match dwallet_mpc_outputs_manager {
             Some(dwallet_mpc_outputs_manager) => Ok(dwallet_mpc_outputs_manager.lock().await),
             None => Err(PeraError::from(
