+++ crates/pera-node/src/lib.rs
@@ -212,8 +212,9 @@ mod simulator {
 use simulator::*;
 
 use pera_core::consensus_handler::ConsensusHandlerInitializer;
+use pera_core::dwallet_mpc::batches_manager::DWalletMPCBatchesManager;
 use pera_core::dwallet_mpc::mpc_manager::{DWalletMPCChannelMessage, DWalletMPCManager};
-use pera_core::dwallet_mpc::mpc_outputs_manager::DWalletMPCOutputsManager;
+use pera_core::dwallet_mpc::mpc_outputs_verifier::DWalletMPCOutputsVerifier;
 use pera_core::safe_client::SafeClientMetricsBase;
 use pera_core::validator_tx_finalizer::ValidatorTxFinalizer;
 use pera_types::execution_config_utils::to_binary_config;
@@ -1296,20 +1297,19 @@ impl PeraNode {
 
         // Start the dWallet MPC manager on epoch start.
         epoch_store
-            .set_dwallet_mpc_outputs_manager(DWalletMPCOutputsManager::new(&epoch_store))
-            .await?;
+            .set_dwallet_mpc_outputs_manager(DWalletMPCOutputsVerifier::new(&epoch_store))?;
 
-        epoch_store
-            .set_dwallet_mpc_sender(
-                DWalletMPCManager::try_new(
-                    Arc::new(consensus_adapter.clone()),
-                    Arc::clone(&epoch_store),
-                    epoch_store.epoch(),
-                    config.clone(),
-                )
-                .await?,
+        epoch_store.set_dwallet_mpc_batches_manager(DWalletMPCBatchesManager::new())?;
+
+        epoch_store.set_dwallet_mpc_sender(
+            DWalletMPCManager::try_new(
+                Arc::new(consensus_adapter.clone()),
+                Arc::clone(&epoch_store),
+                epoch_store.epoch(),
+                config.clone(),
             )
-            .await?;
+            .await?,
+        )?;
 
         let throughput_calculator = Arc::new(ConsensusThroughputCalculator::new(
             None,
